+++ lagom/experimental/exceptions.py
@@ -1,0 +1,7 @@
+from lagom.exceptions import LagomException
+
+
+class DependencyNotDefined(ValueError, LagomException):
+    """The type must be explicitly defined in the container"""
+
+    pass